refactor: Node project much more readable and organized.
