+++ back/node/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "dotenv": "^16.4.5",
     "express": "^4.18.3",
     "http-proxy-middleware": "^2.0.6",
     "jsonwebtoken": "^9.0.2",
